Add logic and authorization to products and categories

+++ src/products/products.controller.ts
@@ -9,6 +9,9 @@ import {
   Query,
   DefaultValuePipe,
   ParseIntPipe,
+  Req,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDto } from './dto/create-product.dto';
@@ -22,13 +25,25 @@ export class ProductsController {
   async create(
     @Param('categoryId') categoryId: number,
     @Body() createProductDto: CreateProductDto,
+    @Req() req,
   ) {
-    const product = await this.productsService.create(
-      createProductDto,
-      +categoryId,
-    );
+    const currentUser = req.currentUserRole;
+    if (currentUser) {
+      const product = await this.productsService.create(
+        createProductDto,
+        +categoryId,
+      );
 
-    return product;
+      return product;
+    } else {
+      throw new HttpException(
+        {
+          status: HttpStatus.UNAUTHORIZED,
+          error: 'ONLY MANAGERS CAN CREATE A NEW PRODUCT',
+        },
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
   }
 
   @Get(':categoryId/products')
@@ -42,18 +57,67 @@ export class ProductsController {
     });
   }
 
+  @Get(':categoryId/disabledProducts')
+  async findDisabled(
+    @Query('page', new DefaultValuePipe(1), ParseIntPipe) page: number,
+    @Query('perPage', new DefaultValuePipe(2), ParseIntPipe) perPage: number,
+    @Req() req,
+  ) {
+    const currentUser = req.currentUserRole;
+    if (currentUser) {
+      return this.productsService.findDisabled({
+        page,
+        perPage,
+      });
+    } else {
+      throw new HttpException(
+        {
+          status: HttpStatus.UNAUTHORIZED,
+          error: 'ONLY MANAGERS CAN SEE THE LIST OF DISABLED PRODUCTS',
+        },
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
+  }
+
   @Get(':categoryId/products/:id')
   findOne(@Param('id') id: string) {
     return this.productsService.findOne(+id);
   }
 
   @Patch(':categoryId/products/:id')
-  update(@Param('id') id: string, @Body() updateProductDto: UpdateProductDto) {
-    return this.productsService.update(+id, updateProductDto);
+  update(
+    @Param('id') id: string,
+    @Body() updateProductDto: UpdateProductDto,
+    @Req() req,
+  ) {
+    const currentUser = req.currentUserRole;
+    if (currentUser) {
+      return this.productsService.update(+id, updateProductDto);
+    } else {
+      throw new HttpException(
+        {
+          status: HttpStatus.UNAUTHORIZED,
+          error: 'ONLY MANAGERS CAN UPDATE A PRODUCT',
+        },
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
   }
 
   @Delete(':categoryId/products/:id')
-  remove(@Param('id') id: string) {
-    return this.productsService.remove(+id);
+  remove(@Param('id') id: string, @Req() req) {
+    const currentUser = req.currentUserRole;
+    if (currentUser) {
+      return this.productsService.remove(+id);
+    } else {
+      throw new HttpException(
+        {
+          status: HttpStatus.UNAUTHORIZED,
+          error: 'ONLY MANAGERS CAN DELETE A PRODUCT',
+        },
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
   }
 }
